Rename abr directory into adaptive
