+++ src/core/adaptive/README.md
