+++ src/core/adaptive/abr_manager.ts
