+++ src/core/adaptive/guess_based_chooser.ts
