+++ src/core/adaptive/index.ts
