+++ src/core/adaptive/network_analyzer.ts
