+++ src/core/adaptive/representation_estimator.ts
