+++ src/core/adaptive/utils/__tests__/ewma.test.ts
