+++ src/core/adaptive/utils/bandwidth_estimator.ts
