+++ src/core/adaptive/utils/ewma.ts
