+++ src/core/adaptive/utils/last_estimate_storage.ts
