+++ src/core/adaptive/utils/pending_requests_store.ts
