+++ src/core/adaptive/utils/representation_score_calculator.ts
