+++ src/core/adaptive/utils/select_optimal_representation.ts
