+++ src/core/stream/adaptation/create_representation_estimator.ts
@@ -28,7 +28,7 @@ import {
   IRepresentationEstimatorPlaybackObservation,
   IRepresentationEstimator,
   IRepresentationEstimatorCallbacks,
-} from "../../abr";
+} from "../../adaptive";
 import { IReadOnlyPlaybackObserver } from "../../api";
 
 /**
